chore: Update ReactDOM import to use react-dom/client
The code changes update the import statement for ReactDOM in the index.jsx file to use the react-dom/client module instead of the react-dom module. This change is made to optimize the loading of the React application.

Note: The recent user commits and recent repository commits are not relevant for generating the commit message in this case.