chore: Update favicon and manifest links
The code changes update the links for the favicon and manifest files in the index.html file. The previous links used the "%PUBLIC_URL%" placeholder, which is no longer necessary since the project has been configured to use absolute paths. This change ensures that the favicon and manifest files are correctly referenced.